add auth (and utils) mod
