+++ src/utils.rs
@@ -1,0 +1,4 @@
+#[cfg(test)]
+pub fn remove_whitespace(s: &str) -> String {
+    s.replace(|c: char| c.is_whitespace(), "")
+}
